requirements.txt
streamlit
opencv-python
pandas
google-cloud-storage
pytz
google-cloud-aiplatform
reportlab
PyPDF2
langchain
faiss-cpu
langchain-google-vertexai
